examples/python-langchain/pyproject.toml
[project]
name = "chakramcp-example-python"
version = "0.1.0"
description = "Example ChakraMCP agent — Python + LangChain + NVIDIA NIM"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-aws>=0.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
agent = "chakramcp_example.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chakramcp_example"]

[tool.uv]
# uv-specific config — uv reads pyproject.toml for dependencies.
# `uv sync` installs everything into .venv.
